Fixed retrieving subjects, in case there are no subjects.
Also: Removed the default value for subjects when creating a user ("Subjects":0) and the default value for tasks when creating a subject ("Tasks":0). There is no need to create nodes in advance, since push() will create a new node in case this is the first push (e.g. if this is the first time we add a subject, push() will create the node Subjects for us since we specified it in the reference path).